pyproject.toml
[project]
name = "lang-chain-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-openai>=1.1.0",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
]
